+++ internal/swagger/docs.go
@@ -2,25 +2,25 @@
 //
 // These endpoints are provided as a proof of concept for CORS issues.
 //
-//     Schemes: http, https
-//     Host: localhost:8080
-//     BasePath: /
-//     Version: 1.0.0
-//     License: MIT http://opensource.org/licenses/MIT
-//     Contact: antonio-alexander<antonio-alexandere@antonio-alexander.dev> http://antonio-alexander.dev
+//  Schemes: http, https
+//  Host: localhost:8080
+//  BasePath: /
+//  Version: 1.0.0
+//  License: MIT http://opensource.org/licenses/MIT
+//  Contact: antonio-alexander<antonio-alexandere@antonio-alexander.dev> http://antonio-alexander.dev
 //
-//     Consumes:
-//     - application/json
+//  Consumes:
+//  - application/json
 //
-//     Produces:
-//     - application/json
+//  Produces:
+//  - application/json
 //
-//     Security:
-//     - basic
+//   Security:
+//   - basic
 //
-//    SecurityDefinitions:
-//    basic:
-//      type: basic
+//   SecurityDefinitions:
+//   basic:
+//     type: basic
 //
 // swagger:meta
 package swagger