+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] -
+## [1.1.0] - 10/31/25
+
+- updated with client to simulate cors
+- fixed issue where swagger wasn't correct for basic authorization
+
+## [1.0.0] - 08/27/22
 
 - Initial release